pyproject.toml
[project]
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.68.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
name = "sample-py-cpp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["E501", "F841", "F401", "F821"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
